Changes FacetRange types
Continuous facet values from the database are f32s, so the FacetRange
should also use f32s. Added Copy as a derived attribute for library
usage.

+++ src/data_structures/facets.rs
@@ -18,8 +18,8 @@ pub enum FacetCoverage {
     Source,
 }
 
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct FacetRange(pub f64, pub f64);
+#[derive(Clone, Copy, Serialize, Deserialize, Debug)]
+pub struct FacetRange(pub f32, pub f32);
 
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct Facet {
